requirements.txt
anthropic
fastapi
uvicorn
chromadb
python-dotenv
tiktoken
pydantic
voyageai
tenacity
slowapi